rename canvas files
