+++ src/brushes/spray_brush.class.ts
@@ -3,7 +3,7 @@ import { Group } from '../shapes/Group';
 import { Shadow } from '../Shadow';
 import { Rect } from '../shapes/Rect';
 import { getRandomInt } from '../util/internals';
-import type { Canvas } from '../canvas/canvas_events';
+import type { Canvas } from '../canvas/CanvasEvents';
 import { BaseBrush } from './base_brush.class';
 
 export type SprayBrushPoint = {
