+++ src/canvas/StaticCanvas.ts
