+++ src/canvas/canvas_gestures.mixin.ts
@@ -6,7 +6,7 @@ import {
   degreesToRadians,
   radiansToDegrees,
 } from '../util/misc/radiansDegreesConversion';
-import { Canvas } from './canvas_events';
+import { Canvas } from './CanvasEvents';
 
 /**
  * Adds support for multi-touch gestures using the Event.js library.
